+++ R/trakt.shows.related.R
@@ -5,6 +5,8 @@
 #' Receive a set of shows that are related to a specific show.
 #' @param target The \code{id} of the show requested. Either the \code{slug}
 #' (e.g. \code{"game-of-thrones"}), \code{trakt id} or \code{IMDb id}
+#' @param extended Whether extended info should be provided.
+#' Defaults to \code{"min"}, can either be \code{"min"} or \code{"full"}
 #' @return A \code{data.frame} containing search results
 #' @export
 #' @note See \href{http://docs.trakt.apiary.io/#reference/shows/related/get-related-shows}{the trakt API docs for further info}
@@ -15,16 +17,15 @@
 #' get_trakt_credentials() # Set required API data/headers
 #' related <- trakt.shows.related("game-of-thrones")
 #' }
-trakt.shows.related <- function(target){
-  ids <- NULL
+trakt.shows.related <- function(target, extended = "min"){
 
   # Construct URL, make API call
-  baseURL  <- "https://api-v2launch.trakt.tv/shows"
-  url      <- paste0(baseURL, "/", target, "/related")
+  baseURL  <- "https://api-v2launch.trakt.tv/shows/"
+  url      <- paste0(baseURL, target, "/related?extended=", extended)
   response <- trakt.api.call(url = url)
 
   # Flattening
-  response <- cbind(subset(response, select = -ids), response$ids)
+  response <- cbind(response[names(response) != "ids"], response$ids)
 
   return(response)
 }